Task 0-Write a function that prints all the elements of a listint_t list. Task 1-Write a function that returns the number of elements in a linked listint_t list. Task 2-Write a function that adds a new node at the beginning of a listint_t list. Task 3-Write a function that adds a new node at the end of a listint_t list.

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,7 +8,7 @@
 size_t print_listint(const listint_t *h)
 {
 	const listint_t *my_print_list;
-	unsigned int i = 0;
+	size_t i = 0;
 
 	my_print_list = malloc(sizeof(listint_t));
 
@@ -28,5 +28,4 @@ size_t print_listint(const listint_t *h)
 	}
 
 	return (i);
-
 }
